requirements.txt
Flask==3.0.0
Flask-Mail==0.10.0
Flask-WTF==1.2.1
pylint==2.11.1
python-dotenv==1.0.1
requests==2.26.0
Werkzeug==3.0.1

gunicorn